moved some of the logic in the app js file to the useMovement hook
