+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"C:\\xampp\\htdocs\\react-apps\\20-react-apps\\react-treasure-hunt-game\\src\\index.js":"1","C:\\xampp\\htdocs\\react-apps\\20-react-apps\\react-treasure-hunt-game\\src\\App.js":"2"},{"size":197,"mtime":1609738360984,"results":"3","hashOfConfig":"4"},{"size":2498,"mtime":1610001564208,"results":"5","hashOfConfig":"4"},{"filePath":"6","messages":"7","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"8"},"185veh5",{"filePath":"9","messages":"10","errorCount":12,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"C:\\xampp\\htdocs\\react-apps\\20-react-apps\\react-treasure-hunt-game\\src\\index.js",[],["11","12"],"C:\\xampp\\htdocs\\react-apps\\20-react-apps\\react-treasure-hunt-game\\src\\App.js",["13","14","15","16","17","18","19","20","21","22","23","24","25"],{"ruleId":"26","replacedBy":"27"},{"ruleId":"28","replacedBy":"29"},{"ruleId":"30","severity":2,"message":"31","line":34,"column":9,"nodeType":"32","messageId":"33","endLine":34,"endColumn":18},{"ruleId":"30","severity":2,"message":"31","line":35,"column":9,"nodeType":"32","messageId":"33","endLine":35,"endColumn":18},{"ruleId":"30","severity":2,"message":"31","line":36,"column":9,"nodeType":"32","messageId":"33","endLine":36,"endColumn":18},{"ruleId":"30","severity":2,"message":"31","line":37,"column":9,"nodeType":"32","messageId":"33","endLine":37,"endColumn":18},{"ruleId":"30","severity":2,"message":"34","line":39,"column":43,"nodeType":"32","messageId":"33","endLine":39,"endColumn":44},{"ruleId":"30","severity":2,"message":"35","line":39,"column":46,"nodeType":"32","messageId":"33","endLine":39,"endColumn":47},{"ruleId":"36","severity":1,"message":"37","line":40,"column":6,"nodeType":"38","endLine":40,"endColumn":12,"suggestions":"39"},{"ruleId":"30","severity":2,"message":"34","line":40,"column":7,"nodeType":"32","messageId":"33","endLine":40,"endColumn":8},{"ruleId":"30","severity":2,"message":"35","line":40,"column":10,"nodeType":"32","messageId":"33","endLine":40,"endColumn":11},{"ruleId":"30","severity":2,"message":"40","line":50,"column":13,"nodeType":"32","messageId":"33","endLine":50,"endColumn":17},{"ruleId":"30","severity":2,"message":"40","line":57,"column":13,"nodeType":"32","messageId":"33","endLine":57,"endColumn":17},{"ruleId":"30","severity":2,"message":"40","line":64,"column":13,"nodeType":"32","messageId":"33","endLine":64,"endColumn":17},{"ruleId":"30","severity":2,"message":"40","line":71,"column":13,"nodeType":"32","messageId":"33","endLine":71,"endColumn":17},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-undef","'direction' is not defined.","Identifier","undef","'x' is not defined.","'y' is not defined.","react-hooks/exhaustive-deps","React Hook useEffect has missing dependencies: 'linkDownRef', 'linkLeftRef', 'linkRightRef', and 'linkUpRef'. Either include them or remove the dependency array. Outer scope values like 'x' aren't valid dependencies because mutating them doesn't re-render the component.","ArrayExpression",["43"],"'move' is not defined.","no-global-assign","no-unsafe-negation",{"desc":"44","fix":"45"},"Update the dependencies array to be: [linkDownRef, linkLeftRef, linkRightRef, linkUpRef]",{"range":"46","text":"47"},[1308,1314],"[linkDownRef, linkLeftRef, linkRightRef, linkUpRef]"]
