+++ src/App.js
@@ -7,9 +7,7 @@ export default function App() {
   const linkUpRef = useState(null);
   const linkLeftRef = useState(null);
   const linkRightRef = useState(null);
-  const [x, setX] = useState(0);
-  const [y, setY] = useState(0);
-  const [direction, setDirection] = useState("down");
+
 
   //set the height and width of the canvas
   //mimics componentDidMount
@@ -41,29 +39,6 @@ export default function App() {
     context.drawImage(theLinkRef.current, x, y);
   }, [x, y]);
 
-  // add event listener to window to listen for arrow keys
-  useEffect(() => {
-    window.addEventListener("keydown", handleKeyDown);
-
-    function handleKeyDown(e) {
-      //console.log(e.keyCode, e);
-      //sconsole.log(e.key, e);
-      if (e.key === "ArrowUp") move("up");
-      if (e.key === "ArrowLeft") move("left");
-      if (e.key === "ArrowDown") move("down");
-      if (e.key === "ArrowRight") move("right");
-    }
-
-    return () => window.removeEventListener("keydown", handleKeyDown);
-  }, []);
-
-  function move(direction) {
-    setDirection(direction);
-    if (direction === "up") setY((y) => y - 20);
-    if (direction === "left") setX((x) => x - 20);
-    if (direction === "down") setY((y) => y + 20);
-    if (direction === "right") setX((x) => x + 20);
-  }
 
   return (
     <div className="app">
